requirements.txt
anyio==3.6.2
autopep8==2.0.2
click==8.1.3
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.6
PyJWT==2.6.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
uvicorn==0.21.1
